I have added a simple interface for my table, changed code in tstStaff.

+++ Testing3/tstStaff.cs
@@ -59,6 +59,7 @@ namespace Testing3
             //create an instance of the class we want to create
             clsStaff AnStaff = new clsStaff();
             //create some test data to assign to the property
+            // DateTime TestData = new DateTime (1999, 7, 10);
             DateTime TestData = DateTime.Now.Date;
             //assign the data to the property
             AnStaff.DOB = TestData;
@@ -83,7 +84,7 @@ namespace Testing3
             //create an instance of the class we want to create
             clsStaff AnStaff = new clsStaff();
             //create some test data to assign to the property
-            Int32 TestData = 1;
+            decimal TestData = 1;
             //assign the data to the property
             AnStaff.MonthlySalary = TestData;
             //test to see that the two values are the same
